wip: some messing with types
