+++ packages/timeline-state-resolver-types/src/expectedPlayoutItems.ts
@@ -1,3 +1,3 @@
-import { VIZMSEPlayoutItemContent } from './integrations/vizMSE.js'
+import { VIZMSEPlayoutItemContent } from './integrations/vizMSE/timeline.js'
 
 export type ExpectedPlayoutItemContent = VIZMSEPlayoutItemContent
