+++ packages/timeline-state-resolver-types/src/index.ts
@@ -2,58 +2,58 @@ import * as Timeline from './superfly-timeline/index.js'
 import { TSRTimelineObjProps } from './mapping.js'
 import { Content } from './superfly-timeline/index.js'
 
-import { TimelineContentTelemetricsAny } from './integrations/telemetrics.js'
-import { TimelineContentAtemAny } from './integrations/atem.js'
-import { TimelineContentCasparCGAny } from './integrations/casparcg.js'
-import { TimelineContentHTTPSendAny } from './integrations/httpSend.js'
-import { TimelineContentTCPSendAny } from './integrations/tcpSend.js'
-import { TimelineContentHyperdeckAny } from './integrations/hyperdeck.js'
-import { TimelineContentLawoAny } from './integrations/lawo.js'
-import { TimelineContentOSCAny } from './integrations/osc.js'
-import { TimelineContentPharosAny } from './integrations/pharos.js'
-import { TimelineContentPanasonicPtzAny } from './integrations/panasonicPTZ.js'
-import { TimelineContentAbstractAny } from './integrations/abstract.js'
-import { TimelineContentQuantelAny } from './integrations/quantel.js'
-import { TimelineContentShotoku } from './integrations/shotoku.js'
-import { TimelineContentSisyfosAny } from './integrations/sisyfos.js'
-import { TimelineContentSofieChefAny } from './integrations/sofieChef.js'
-import { TimelineContentVIZMSEAny } from './integrations/vizMSE.js'
-import { TimelineContentSingularLiveAny } from './integrations/singularLive.js'
-import { TimelineContentVMixAny } from './integrations/vmix.js'
-import { TimelineContentOBSAny } from './integrations/obs.js'
-import { TimelineContentTriCasterAny } from './integrations/tricaster.js'
-import { TimelineContentWebSocketClientAny } from './integrations/websocketClient.js'
-import { TimelineContentKairosAny } from './integrations/kairos.js'
-import { DeviceType } from './generated/index.js'
-import { TimelineContentUDPSendAny } from './integrations/udpSend.js'
-import { TimelineContentOgrafAny } from './integrations/ograf.js'
+import { TimelineContentTelemetricsAny } from './integrations/telemetrics/timeline.js'
+import { TimelineContentAtemAny } from './integrations/atem/timeline.js'
+import { TimelineContentCasparCGAny } from './integrations/casparcg/timeline.js'
+import { TimelineContentHTTPSendAny } from './integrations/httpSend/timeline.js'
+import { TimelineContentTCPSendAny } from './integrations/tcpSend/timeline.js'
+import { TimelineContentHyperdeckAny } from './integrations/hyperdeck/timeline.js'
+import { TimelineContentLawoAny } from './integrations/lawo/timeline.js'
+import { TimelineContentOSCAny } from './integrations/osc/timeline.js'
+import { TimelineContentPharosAny } from './integrations/pharos/timeline.js'
+import { TimelineContentPanasonicPtzAny } from './integrations/panasonicPTZ/timeline.js'
+import { TimelineContentAbstractAny } from './integrations/abstract/timeline.js'
+import { TimelineContentQuantelAny } from './integrations/quantel/timeline.js'
+import { TimelineContentShotoku } from './integrations/shotoku/timeline.js'
+import { TimelineContentSisyfosAny } from './integrations/sisyfos/timeline.js'
+import { TimelineContentSofieChefAny } from './integrations/sofieChef/timeline.js'
+import { TimelineContentVIZMSEAny } from './integrations/vizMSE/timeline.js'
+import { TimelineContentSingularLiveAny } from './integrations/singularLive/timeline.js'
+import { TimelineContentVMixAny } from './integrations/vmix/timeline.js'
+import { TimelineContentOBSAny } from './integrations/obs/timeline.js'
+import { TimelineContentTriCasterAny } from './integrations/tricaster/timeline.js'
+import { TimelineContentWebSocketClientAny } from './integrations/websocketClient/timeline.js'
+import { TimelineContentKairosAny } from './integrations/kairos/timeline.js'
+import { DeviceType, TSRDeviceTypesMap } from './generated/index.js'
+import { TimelineContentUDPSendAny } from './integrations/udpSend/timeline.js'
+import { TimelineContentOgrafAny } from './integrations/ograf/timeline.js'
 
-export * from './integrations/abstract.js'
-export * from './integrations/atem.js'
-export * from './integrations/casparcg.js'
-export * from './integrations/httpSend.js'
-export * from './integrations/hyperdeck.js'
-export * from './integrations/kairos.js'
-export * from './integrations/lawo.js'
-export * from './integrations/ograf.js'
-export * from './integrations/osc.js'
-export * from './integrations/pharos.js'
-export * from './integrations/panasonicPTZ.js'
-export * from './integrations/sisyfos.js'
-export * from './integrations/sofieChef.js'
-export * from './integrations/quantel.js'
-export * from './integrations/shotoku.js'
-export * from './integrations/tcpSend.js'
-export * from './integrations/vizMSE.js'
-export * from './integrations/singularLive.js'
-export * from './integrations/vmix.js'
-export * from './integrations/obs.js'
-export * from './integrations/tricaster.js'
-export * from './integrations/telemetrics.js'
-export * from './integrations/multiOsc.js'
-export * from './integrations/udpSend.js'
-export * from './integrations/viscaOverIP.js'
-export * from './integrations/websocketClient.js'
+export * from './integrations/abstract/timeline.js'
+export * from './integrations/atem/timeline.js'
+export * from './integrations/casparcg/timeline.js'
+export * from './integrations/httpSend/timeline.js'
+export * from './integrations/hyperdeck/timeline.js'
+export * from './integrations/kairos/timeline.js'
+export * from './integrations/lawo/timeline.js'
+export * from './integrations/ograf/timeline.js'
+export * from './integrations/osc/timeline.js'
+export * from './integrations/pharos/timeline.js'
+export * from './integrations/panasonicPTZ/timeline.js'
+export * from './integrations/sisyfos/timeline.js'
+export * from './integrations/sofieChef/timeline.js'
+export * from './integrations/quantel/timeline.js'
+export * from './integrations/shotoku/timeline.js'
+export * from './integrations/tcpSend/timeline.js'
+export * from './integrations/vizMSE/timeline.js'
+export * from './integrations/singularLive/timeline.js'
+export * from './integrations/vmix/timeline.js'
+export * from './integrations/obs/timeline.js'
+export * from './integrations/tricaster/timeline.js'
+export * from './integrations/telemetrics/timeline.js'
+export * from './integrations/multiOsc/timeline.js'
+export * from './integrations/udpSend/timeline.js'
+export * from './integrations/viscaOverIP/timeline.js'
+export * from './integrations/websocketClient/timeline.js'
 
 export * from './actions.js'
 export * from './datastore.js'
@@ -63,8 +63,8 @@ export * from './mapping.js'
 export * from './mediaObject.js'
 export * from './templateString.js'
 export * from './translations.js'
-
 export * from './generated/index.js'
+
 export { Timeline }
 
 export interface TSRTimelineKeyframe<TContent> extends Omit<Timeline.TimelineKeyframe, 'content'> {
@@ -137,6 +137,8 @@ export interface TimelineContentMap {
 }
 
 export type TSRTimelineContent = TimelineContentMap[keyof TimelineContentMap]
+
+export type TSRMappingOptions = TSRDeviceTypesMap[keyof TSRDeviceTypesMap]['Mappings']
 
 /**
  * A simple key value store that can be referred to from the timeline objects
