--- packages/timeline-state-resolver-types/src/integrations/abstract.ts
@@ -1,7 +1,0 @@
-import { DeviceType } from '../generated/index.js'
-
-export type TimelineContentAbstractAny = TSRTimelineContentAbstract
-export interface TSRTimelineContentAbstract {
-	deviceType: DeviceType.ABSTRACT
-	//		[key: string]: any
-}
