+++ packages/timeline-state-resolver-types/src/integrations/abstract/events.ts
@@ -1,0 +1,5 @@
+export type AbstractEvents = {
+	something: {
+		abc: string
+	}
+}
