+++ packages/timeline-state-resolver-types/src/integrations/abstract/timeline.ts
@@ -1,0 +1,7 @@
+import { DeviceType } from '../../generated/index.js'
+
+export type TimelineContentAbstractAny = TSRTimelineContentAbstract
+export interface TSRTimelineContentAbstract {
+	deviceType: DeviceType.ABSTRACT
+	//		[key: string]: any
+}
