+++ packages/timeline-state-resolver-types/src/integrations/atem/events.ts
@@ -1,0 +1,8 @@
+export type AtemEvents = {
+	[key: `me.${number}.test`]: {
+		abc: string
+	}
+	[key: `me.${number}.another`]: {
+		def: string
+	}
+}
