+++ packages/timeline-state-resolver-types/src/integrations/multiOsc/timeline.ts
