+++ packages/timeline-state-resolver-types/src/integrations/sofieChef/timeline.ts
@@ -1,0 +1,17 @@
+import { DeviceType, TemplateString } from '../../index.js'
+
+export enum TimelineContentTypeSofieChef {
+	URL = 'url',
+}
+
+export type TimelineContentSofieChefAny = TimelineContentSofieChefScene
+
+export interface TimelineContentSofieChef {
+	deviceType: DeviceType.SOFIE_CHEF
+	type: TimelineContentTypeSofieChef
+}
+export interface TimelineContentSofieChefScene extends TimelineContentSofieChef {
+	type: TimelineContentTypeSofieChef.URL
+
+	url: string | TemplateString
+}
