--- packages/timeline-state-resolver-types/src/integrations/telemetrics.ts
@@ -1,8 +1,0 @@
-import { DeviceType } from '../index.js'
-
-export type TimelineContentTelemetricsAny = TimelineContentTelemetrics
-
-export interface TimelineContentTelemetrics {
-	deviceType: DeviceType.TELEMETRICS
-	presetShotIdentifiers: number[]
-}
