+++ packages/timeline-state-resolver-types/src/integrations/telemetrics/timeline.ts
@@ -1,0 +1,8 @@
+import { DeviceType } from '../../index.js'
+
+export type TimelineContentTelemetricsAny = TimelineContentTelemetrics
+
+export interface TimelineContentTelemetrics {
+	deviceType: DeviceType.TELEMETRICS
+	presetShotIdentifiers: number[]
+}
