--- packages/timeline-state-resolver-types/src/integrations/viscaOverIP.ts
@@ -1,8 +1,0 @@
-import { DeviceType } from '../generated/index.js'
-
-export type TimelineContentViscaOverIpAny = TimelineContentViscaOverIp
-export interface TimelineContentViscaOverIpBase {
-	deviceType: DeviceType.VISCA_OVER_IP
-}
-
-export type TimelineContentViscaOverIp = TimelineContentViscaOverIpBase
