+++ packages/timeline-state-resolver-types/src/integrations/viscaOverIP/timeline.ts
@@ -1,0 +1,8 @@
+import { DeviceType } from '../../generated/index.js'
+
+export type TimelineContentViscaOverIpAny = TimelineContentViscaOverIp
+export interface TimelineContentViscaOverIpBase {
+	deviceType: DeviceType.VISCA_OVER_IP
+}
+
+export type TimelineContentViscaOverIp = TimelineContentViscaOverIpBase
